#20 Use Profile Manager
Switch to using the pac4j ProfileManager.
This also includes a rewrite of vertx-pac4j to make it follow the patterns established
in play-pac4j and j2e-pac4j.

Note that there is some outstanding work to incorporate the latest changes made to
play-pac4j, and also to use the full pac4j authorization capability (the current codebase
uses the pac4j-identified authorities but following the vert.x model which is more
restrictive than the pac4j model). I plan to resolve both of these issues in subsequent
pull requests and will raise issues for both.