Issue #40 Use the authorization checker functionality in Pac4j for authorization
We've now amended the vertx-pac4j implementation to use the newly-introduced
client finder functionality to determine the client to use.

The second part of this issue relates to using the pac4j 1.8 authorization checker
functionality. Override the current authorise method so that it will use the pac4j auth
checkers and amend tests accordingly.

Note that we've chosen to override the authorise function inherited from the vert.x base
auth handler class. This is so that if someone familiar with vert.x wishes to use their own
mechanism for authorization, bypassing entirely the Pac4j authorization infrastructure,
they can override the standard method exposed by the vert.x handler for doing so. I'm
not sure why someone would choose to do that, but this way they have the option to do
so, or to write their own Pac4j authorizer. Either approach will work.

Note that this will also resolve issue #23 - use Pac4j authorization