Issue #48 RequiresAuthenticationHandler (and derivatives) should not require an authorizer name to be set
Previously RequiresAuthenticationHandler and its derivatives mandate that an authorizer name be supplied on instantiation (via the Pac4jAuthHandlerOptions constructor). This is inconsistent with the play-pac4j implementation and more critically means that if an authorizer is required for any endpoint protected by Pac4j, an authorizer is required for all endpoints.

Modify the RequiresAuthenticationHandler and its derivatives to ensure that authorizer name and client name are not necessary and default to empty string (as per play-pac4j implementation). Where authorizer name is not supplied, then authorization should succeed on that endpoint for all authenticated users.