Issue #52 DefaultHttpActionAdapter should handle failure scenarios (specifically 401/403) by failing the RoutingContext
The DefaultHttpActionAdapter for vert.x handled the authentication/authorization failure scenarios incorrectly from a vert.x point of view.In the event of a 401/403 failure we would sett the status code, write out a default response body (which noone is likely to want to use) and end the response normally.

This essentially made it impossible for someone using vertx-pac4j to customise those error responses, meaning that the default response bodies would always be received. The correct approach from a vert.x-web idiom is to add a failure handler to all routes which checks the status code and delivers an appropriate (in the context of the application) response to the client. However, because we were ending the response normally, no failure handler would be called.

Therefore for failure, we have modified so that we ultimately call RoutingContext::fail. This required a modification to VertxWebContext to expose a fail method wrapping RoutingContext::fail) which will trigger the failure handler with the relevant status code, and then the vertx-pac4j user should provide a failure handler which delivers appropriate content for 401/403 failures (and any other failure scenarios for which they wish to deliver a response body.