Issue #54 Create a CAS logout handler
Create a CAS logout handler comparable to that used for play-pac4j.

Two CAS logout handler implementations will be provided for vert.x, based on using vert.x shared data as a cache for mapping a CAS ticket to a session (to enable single sign-out).

The first is for a non-clustered vert.x deployment, and is based on a local cache (ie within the same JVM).

The second is for a clustered vert.x deployment and is based on a clustered shared data (i.e. clustered across multiple JVMs).

Note that for the clustered deployment, because of the blocking behaviour of Pac4j under the hood, we actually have to take non-blocking API calls and wrap them in a blocking API, which is not ideal but unfortunately necessary. When we start to work on a non-blocking version of Pac4j, we should address this by making the CAS logout handler an async API.

Note also that developers are perfectly welcome to write their own implementations based on other caches if they so choose. The vert.x versions are provided as a courtesy and a convenience.