Issue #55 Full fix, including the following elements which weren't entirely dealt with by the original PR
(1) Ensure fix works for both direct and indirect clients
(2) Ensure that where we're using a UserSessionHandler (which auto-writes vertx user to session) the vertx user stored properly supports clustered sessions
(3) Ensure that external logout trigger (e.g. via Cas logout handler) triggers removal of the vertx user as well as Pac4j user - otherwise Cas logout requests from Cas server will not achieve desired result. Work for this trigger is currently WIP and expected to be complete before I start on this fix, if I shelve that work to work on this, I must add these notes to the Cas logout handler.

Note that due to changes in the VertxWebContext it became necessary to use mocks for a couple of the tests, in general I'd prefer to avoid mocks, but inheriting/stubbing a real implementation would be more painful than use of mock-based stubs in this case, so I preferred the less painful/awkward option.