Issue #60 Add WebContext Matcher support
In pac4j 1.8.1, the concept of a matcher was added - this allows definition of filters for webcontexts. If a webcontext matches the filter, apply the pac4j functionality, otherwise assume we are authorized.

Add support (and tests) for this web context matching mechanism. An example of this use case might be to exclude specific sub-paths from the pac4j handling for a given handler for instance.