Cluster serializing a Pac4jUser via writeToBuffer then deserializing via readFromBuffer fails with the following error.
This occurs because there are two errors in the serialization/deserialization loop. Firstly I was using the principal as a convenience for serialization, however, we may wish to restrict the attributes visible in the principal, so it is, in my view, better to serialize the actual UserProfile instance to a Json string and write that to the buffer.

Secondly when deserializing, it's necessary to convert the Json string we read from the buffer to a JsonObject, and then apply the Json decoding. Implemented this and validated with a round trip test