Issue #70 Update to pac4j 1.9.1
Further changes in response to PR comments.

Also added a test for the ApplicationLogoutHandler behaviour (default and as exercised by vertx-pac4j-demo) so that when we migrate to DefaultApplicationLogoutLogic we can assert the expected behaviour.

Note that it appears that for pac4j 1.9.1 we can't migrate to DefaultApplicationLogoutLogic as it doesn't support custom profile manager creation (either via overriding a method or via injection of a factory function)