Issue #74 Update to pac4j 1.9.2 which puts vertx-pac4j onto latest releases of both vert.x and pac4j.
Retire the VertxSecurityLogic and VertxCallback Logic classes by injecting profile manager factory lambdas into the Default logics.

Note that this commit does not update ApplicationLogoutHandler to use the relevant Default logic yet, this is to be added shortly.

Note also that there was an API compatibility break between pac4j 1.9.1 and 1.9.2 due to relocation of an enumeration, so small changes had to be made around this.