Issue #72 Implement multi-profile support in vertx-pac4j
Multi-profile support in vertx-pac4j implies the need for a map of user profiles within the Pac4jUser instead of just one user profile.

This has implications for serialization for clustered sessions, so this commit will, if applied, make multi-profile support possible within the Pac4j user and also provide a test for serialization/deserialization using existing mechanisms to make sure clustered sessions will adequately be supported.

It will not provide full implementation of multi-profile behaviour in the handler or tests for this - these will be added in a subsequent commit, but this commit is a necessary building block for multi profile support.