Make the MultiProfileTest compile-time check for the assertThat method rather than runtime
We had a run-time check in on this interface to check the superclass. This doesn't really make sense as we could just as easily make the required method part of the interface and then refer to it in our default implementations. If we find a more generic case, then we can have this interface extend another interface which exposes the relevant required method.