#82 Ensure tests route log output to slf4j
vert.x does not output logs to slf4j by default but uses JUL for its logging. If a user/developer wants to use slf4j logging it is for them to configure this (via a system property and import of appropriate dependencies)

In keeping with the vert.x standards, the vertx-pac4j code follows the same standard (but can be made via config to route to slf4j). This commit, will, if merged, ensure that for the tests, when run through maven, all the standard vertx logging is routed to slf4j and can therefore be reviewed in tandem with the slf4j logging generated within pac4j.